Fixes and cleanup
Our unsubscribe page uses this script, and the page was breaking because "p4_fb_page" was undefined, making the `attr` function throw an error which wasn't caught, interrupting script execution. This change inserts guards to avoid using `attr` with an undefined value. I also took the occasion to do a bit of clean up, reducing the amount of work javascript has to do as well as the number of lines. I also found places using `event.preventDefault()` or `event.stopPropagation()` without naming the first argument to the function, effectively making this `undefined.preventDefault()` or `undefined.stopPropagation(). This change also fixes that.